Sound! After fumbling a bit to get the sound library wired in properly, sound worked with only small bugs in the apu. Which is shocking. More on bugs later.
This required a big architectrual change to.
the core loop. It operates pretty differently
now.

Before the core was running the chips long enough
to generate one video frame, blitting
that out, then sleeping to wait for the next
frame.

Now the main driver is the audio library.
The audio thread pulls samples from a shared
queue at the rate the hardware needs them.
At the same time, the main thread clocks
the chips as fast as it can, throwing
audio samples into that queue. When
the queue is full, it blocks the main
thread from advancing until the audio hardware
is ready for more data and pulls from the queue.
Hence the audio hardware/driver/etc is driving
the timing of the emulation now. Video
frames are rendered whenever a full frame is
ready in the buffer, but that timing can
be "unrelated" to audio timing.

Bugs:
1. When SMB first starts there's an
audible "pop." It might be that the APU
needs to do like the PPU does and ignore
writes to its registers for a period on
startup/reset. Haven't dealt with this yet.
May be a legit artifact of the game?

2. Some sounds were lasting too long or
even "forever" until another sound on the
same channel canceled. I had forgotten
to clear the restart flag on the
sweeper unit so it was forever resetting
back to the beginning and not actually
sweeping. Fixed.