Added in DMC DMA timing and DMC IRQs.
Along the way I realized the PPU DMA
timing was all wrong because I was using
APU cycle for all the alignment, but
actually all DMA has to align with the
CPU first then the APU. That's fixed now.
While I was at it I turned those cycle
alignments into enums to reduce the
boolean blindness.