Implemented 6 more easy mappers.
I've set out to do the 10 most
commonly used mappers. So I've
added 4 easy out of that list to
get a feel for how mappers work
in addition to the 1 I already had.
I also threw in 2 more mappers that
aren't top 10 but were easy
variations on ones that I had built.

When I started I realized that
mapper internal address spaces
could be much larger than the
external address space - that's
one of the main points of mappers!
So I had to change data types and
get rid of masking limits.