Refactoring Cartridge Part 1
For some mappers it turns out that
it would be easier if the mapper
intercepted reads and writes directly
rather than the protocol I have now.
As a first step I've moved all the
info lifted from the NES header
into a struct that is passed to
mappers for initialization.