Bugs with NMI and sprites.
Added suppression of VBL flag when
it would be raised on the same
tick when it's lowered. Fixed
an issue where sprite 0 collision
was being detected even when sprites
or bg were disabled. Tweaked the
timing of determining nmi status
in order to get a bit closer to
proper timing.